chore: 기본 폴더구조 수정 및 도메인 로직 분리
