+++ README.md
@@ -15,7 +15,7 @@
   - [x] 모달
   - [ ] 검색 리스트
     - [x] 단일 검색 리스트 요소
-    - [ ] 최근 검색어
+    - [x] 최근 검색어
     - [x] 추천 검색어
   - [ ] 헤더
