+++ src/App.tsx
@@ -1,7 +1,16 @@
-import "./App.css";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import MainPage from "@/domain/KoreaClinicalInfo/pages/MainPage";
+import FavoritesPage from "@/domain/KoreaClinicalInfo/pages/FavoritesPage";
 
 function App() {
-  return <>졸려요</>;
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<MainPage />} />
+        <Route path="/favorites" element={<FavoritesPage />} />
+      </Routes>
+    </Router>
+  );
 }
 
 export default App;
