+++ src/domain/KoreaClinicalInfo/components/SearchBar.tsx
