+++ src/domain/KoreaClinicalInfo/components/SearchCard.tsx
