+++ src/domain/KoreaClinicalInfo/components/SearchSuggestion/RecentSearch.tsx
