+++ src/domain/KoreaClinicalInfo/components/SearchSuggestion/RecommendSearch.tsx
