+++ src/domain/KoreaClinicalInfo/components/SearchSuggestion/SearchItem.tsx
