+++ src/domain/KoreaClinicalInfo/components/SearchSuggestion/index.tsx
