+++ src/domain/KoreaClinicalInfo/pages/FavoritesPage.tsx
@@ -1,0 +1,9 @@
+const FavoritesPage = () => {
+  return (
+    <div>
+      <h1>Favorites</h1>
+    </div>
+  );
+};
+
+export default FavoritesPage;
