+++ src/domain/KoreaClinicalInfo/pages/MainPage.tsx
@@ -1,0 +1,9 @@
+const MainPage = () => {
+  return (
+    <div>
+      <h1>메인 페이지</h1>
+    </div>
+  );
+};
+
+export default MainPage;
